requirements.txt
torch==2.8.0
torchvision==0.23.0
Flask==3.0.0
numpy==2.1.0
Pillow==11.3.0
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.0
gunicorn==23.0.0